InputField in data_utils
